Refactor 'when' and 'unless' to types of map
Treat ApplyWhen and ApplyUnless as syntactic sugar for ApplyMap.
